moved wiki score indicator to createWikiTopBar folder
